+++ packages/blocks/core/js/wordpress/navigation-link/index.js
@@ -45,7 +45,7 @@ export const NavigationLink: BlockType = {
 					<h5>{__('Is Current Page?', 'blockera')}</h5>
 					<p>
 						{__(
-							'Activates if the block URL is the same as the current page URL.',
+							'Activates if the block URL is the same as the current page.',
 							'blockera'
 						)}
 					</p>
@@ -58,6 +58,58 @@ export const NavigationLink: BlockType = {
 			breakpoints: {},
 			priority: 10,
 			force: true,
+			settings: {
+				color: 'var(--blockera-controls-states-color)',
+			},
+		},
+		'current-menu-parent': {
+			type: 'current-menu-parent',
+			label: __('Current Page Parent', 'blockera'),
+			category: 'special',
+			tooltip: (
+				<>
+					<h5>{__('Is Current Page Parent?', 'blockera')}</h5>
+					<p>
+						{__(
+							'Indicates that one of this item’s direct descendants is the current page.',
+							'blockera'
+						)}
+					</p>
+					<code style={{ margin: '5px 0' }}>
+						<span style={{ opacity: '0.7' }}>.block</span>
+						.current-menu-parent
+					</code>
+				</>
+			),
+			breakpoints: {},
+			priority: 10,
+			force: false,
+			settings: {
+				color: 'var(--blockera-controls-states-color)',
+			},
+		},
+		'current-menu-ancestor': {
+			type: 'current-menu-ancestor',
+			label: __('Current Page Ancestor', 'blockera'),
+			category: 'special',
+			tooltip: (
+				<>
+					<h5>{__('Is Current Page Ancestor?', 'blockera')}</h5>
+					<p>
+						{__(
+							'Indicates that one of this item’s descendants is the current page.',
+							'blockera'
+						)}
+					</p>
+					<code style={{ margin: '5px 0' }}>
+						<span style={{ opacity: '0.7' }}>.block</span>
+						.current-menu-ancestor
+					</code>
+				</>
+			),
+			breakpoints: {},
+			priority: 10,
+			force: false,
 			settings: {
 				color: 'var(--blockera-controls-states-color)',
 			},